Use the subtitle parameter of the organizer Distribute for Enterprise screen as the equivalent for CFBundleShortVersionString of info.plist
If CFBundleVersion is used as a build number, then the application version, as it will be shown in the AppStore is not visible. In that case the version number usually is placed into CFBundleShortVersionString in info.plist. To show this on the website and the in-app-update dialog, the subtitle parameter of the distribution plist file is reused to store the same information while maintining version number check compatibility. The subtitle value is now exclusively used for this.