Server: Add json API response format version 2
If the optional parameter API is provided, the client can request the json respond for available apps to be in version 2, which means the json result is actually an array, so multiple version information can be returned.
The server currently does not support providing multiple different versions though, so only one element is returned.
Version 1, meaning not enclosing it into an array, is the default