iOS Client Updates
- Fix crash in PSStoreButton
self.superview was used to calculate button frame before it was actually added to a superview, crash in [redacted] beta
- Enhance version view
Add version header, showing version number, date and installed information. Also adjusted font size
- Added convenience methods to BWApp
Version and Date string are required in several places, so rather provide a method to implement it only once
- Added button states
The view controller stores its own button states, and shows them correctly depending on available updates or not.
- Improved rotation rendering
Uses 1 pass rotation method to render, and also set the cell text to be empty, so the view looks better. There is still a problem left when rotating from landscape to portrait, font sizes are weird.
- Adjust HTTP User-Agent to be Hockey/iOS
- Change bundle identifier
Changing to de.buzzworks.hockeydemo and provide demo files on the server