Improve iOS client settings UI
iPhone/iPod: Use a page curl effect to show the settings. On iOS < 4.x do a simple fullscreen modal sheet.
iPad: Use the popover for the settings

Add settings to allow the user to deny sending user and usage data. Only shows if the developer requests that data and he allows the user to overwrite explicitly.