Fix the language detection once more
We don't actually want the system settings language, since that language might now be present in the app. We are rather interested in the apps localization which is being used. Having the name of the corresponding .lproj folder is fine, since we can match that to the real files easily.