Removed Reachability
The implementation expected to find Apples default implementation which made it crash on different mplementations.
Instead of expecting reachability, Hockey now listens to NetworkDidBecomeReachable notifications and will check for updates or authorization.

You have to trigger this notification in your own apps code if you are using a reachability framework.