make HockeyLib static library work in projects with deployment target < 5.0
If HockeyLib is built as a static library and
linked into the target application, the deployment
target of the static lib should not be used to
statically decide which JSON serialization should be used, as
the target application might use a different one. Thus
HOCKEYLIB_STATIC_LIBRARY is defined in the library target
and depending on that define, a compile time decision is made
or not.